requirements.txt
crewai
openai
faiss-cpu
sentence-transformers
langchain-community
requests
